Rename pymepix class to solve name conflict with pymepix package.
